pyproject.toml
[tool.poetry]
name = "lec_ditto"
version = "0.1.0"
description = ""
authors = ["victoravtr <victoravtr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
deepdiff = "^5.7.0"
requests = "^2.26.0"
python-dotenv = "^0.19.2"
ratelimit = "^2.2.1"
telegram-send = "^0.25"
telegram = "^0.0.1"
python-telegram-bot = "^13.10"
telebot = "^0.0.4"
Telethon = "^1.24.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
